webscraping/ingest/finra/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-default-group==1.2.4
filelock==3.19.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
requests-file==2.1.0
rich==14.1.0
rsa==4.9.1
selectolax==0.4.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlite-fts4==1.0.3
sqlite-utils==3.38
tabulate==0.9.0
tenacity==9.1.2
tldextract==5.3.0
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1

httpx[http2]>=0.27.0
selectolax>=0.3.21
beautifulsoup4>=4.12.3
lxml>=5.2.2
tldextract>=5.1.2
pydantic>=2.7.1
pyyaml>=6.0.1
tenacity>=8.3.0
tqdm>=4.66.4
typer>=0.12.3
google-cloud-storage>=2.16.0
python-dotenv>=1.0.1
sqlite-utils>=3.36
orjson>=3.10.7
